+++ pkgs/test_core/lib/src/runner/loader.dart
@@ -19,6 +19,7 @@ import 'package:yaml/yaml.dart';
 import '../util/io.dart';
 import 'compiler_selection.dart';
 import 'configuration.dart';
+import 'global_setup.dart';
 import 'hack_register_platform.dart';
 import 'load_exception.dart';
 import 'load_suite.dart';
@@ -125,6 +126,9 @@ class Loader {
 
   /// The test runner configuration.
   final _config = Configuration.current;
+
+  /// Manages global setup hooks.
+  late final GlobalSetupManager globalSetupManager = GlobalSetupManager();
 
   /// All suites that have been created by the loader.
   final _suites = <RunnerSuite>{};
@@ -383,6 +387,7 @@ class Loader {
 
   /// Closes the loader and releases all resources allocated by it.
   Future close() => _closeMemo.runOnce(() async {
+    await globalSetupManager.close();
     await Future.wait([
       Future.wait(
         _platformPlugins.values.map((memo) async {